Added the ProxyServiceClient to invoke remote API's. In doing this I made the assumption that the REST Proxy will respond to the internal clients via a 'ProxyResponse.java' object, containing the results of the Remote 3rd part API call. A few issues to consider here. Http Headers can contain duplicates, which make this area overly complex. Another area of concern is associated with Synch. Service Timeout. You don't want to wait forever, you may also need to account for 3rd party performance. So it seemed sensible to provide a timeout value per service. The issue here is that this timeout can be one of 3 different points along the invocation process. I've opted to set the timeout, waiting for the remote service to start streaming bytes back along the wire. Not attempting to make the initial connection. Historically setting a webservice timeout in Spring is fiddly. I chose to leverage Apache HTTP Client as a dependancy to the project.
I did consider creating a test for the timeout handling of Remote 3rd part
API, but that was to tricky outside of my setup. Where I have a number of
REST CRUD apis to tinker with.

Additional work is required to unmarshell Headers and QParams from the
ProxyRequest JSON and the MongoDb. I believe the injected JSON and POJO
code needs a small change to make that work correctly.